rename atom folder to state
